+++ ZeroLibraries/Zilch/Project/Zilch/WebSocket.cpp
@@ -82,13 +82,13 @@ namespace Zilch
     }
 
     // Send the header off first (this will block until it sends the entire thing!)
-    int sentSizeHeader = this->RemoteSocket.Send(status, header, (int)headerSize, SocketFlags::None);
+    size_t sentSizeHeader = this->RemoteSocket.Send(status, header, (int)headerSize, SocketFlags::None);
     ErrorIf(sentSizeHeader != 0 && sentSizeHeader != (int)headerSize, "We should always send the entire header, send should block until its all sent!");
     if (status.Failed())
       return;
 
     // Now send the rest of the payload data (could be done in one send, no problem though!)
-    int sentSizeData = this->RemoteSocket.Send(status, data, (int)length, SocketFlags::None);
+    size_t sentSizeData = this->RemoteSocket.Send(status, data, (int)length, SocketFlags::None);
     ErrorIf(sentSizeData != 0 && sentSizeData != (int)length, "We should always send the entire data, send should block until its all sent!");
     if (status.Failed())
       return;
@@ -124,7 +124,7 @@ namespace Zilch
 
       // Read the data from the socket
       byte buffer[ReceiveBufferSize];
-      int amountReceived = this->RemoteSocket.Receive(status, buffer, ReceiveBufferSize, SocketFlags::None);
+      size_t amountReceived = this->RemoteSocket.Receive(status, buffer, ReceiveBufferSize, SocketFlags::None);
       
       // If the receive call failed, or we gracefully disconnected...
       if (status.Failed() || amountReceived == 0)
@@ -256,7 +256,7 @@ namespace Zilch
       return;
 
     // Now listen on the socket, which should allow incoming connections to be accepted
-    this->ListenerSocket.Listen(status, Socket::GetMaxListenBacklog());
+    this->ListenerSocket.Listen(status, static_cast<uint>(Socket::GetMaxListenBacklog()));
   }
   
   //***************************************************************************
@@ -297,7 +297,7 @@ namespace Zilch
     {
       // Read the data from the socket
       byte buffer[ReceiveBufferSize];
-      int amountReceived = connectionOut.RemoteSocket.Receive(status, buffer, ReceiveBufferSize, SocketFlags::None);
+      size_t amountReceived = connectionOut.RemoteSocket.Receive(status, buffer, ReceiveBufferSize, SocketFlags::None);
       
       // If the receive call failed, or we gracefully disconnected...
       if (status.Failed() || amountReceived == 0)
